+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 13.65.0
  - Issue #567 - Show Code Analyzer Rules: Does not pick up multiple custom code analyzers in multi-root workspace projects - case sensitive check fix
  - Issue #586 - Code Cleanup: Sort Usings
+ - Fix Identifiers case command incorrectly detected object IDs
 
 Thank you
  - rvanbekkum for reporting issue #567